requirements.txt
firebase-admin==4.4.0
Flask==1.1.1
flask-accepts==0.16.5
flask-admin==1.5.6
flask-allows==0.7.1
Flask-BabelEx==0.9.4
flask-cors==3.0.8
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
google-api-python-client==1.8.0
google-auth==1.20.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-tasks==1.5.0
google-python-cloud-debugger; sys_platform == 'linux'
html2text==2020.1.16
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
mypy-extensions==0.4.3
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2020.4
requests==2.23.0
SQLAlchemy==1.3.23
Werkzeug==0.16.1
python-magic==0.4.18
